An unnecessary float typecast was removed from the Dijkstra section of RunSearch() in AI.cs.  RunSearch() in DijkstraMatthew.cs was cleaned, and stopwatch functionality was added.

+++ Maze/Maze/AI.cs
@@ -113,7 +113,7 @@ namespace Maze
             Search s = new DijkstraMatthew();
             s.SetupSearch(nodeMap, XStart, YStart, XEnd, YEnd);
             time = (int)s.RunSearch();
-            return (float)time;
+            return time;
          }
          return 0;
       }
